Add MethodNameStartingWithUppercaseLetter cop.
This commit is quite a significant one, marking a key strategical shift as well as some technological changes.

First and foremost, I am now adding cops without autocorrect support, in order to record the differences between Ruby and Crystal, as well as provide a foundation of detection if it becomes possible to autocorrect these offenses.
The goal of this project to automatically convert between Ruby and Crystal remains unchanged, and the reporting of issues that require manual correction is a complement to this, with an eye on eventally finding a way to make them automatic.

Secondly, RSpec is also being used for testing. RSpec will, at least for now, only be used on cops without autocorrect support, as the only validation of their behavior that is required is via such testing.
(As opposed to cops that do support autocorrection, where the main goal is to ensure they produce functional Crystal code-- the best way to test that is through the current setup.)
In the future, I may add RSpec to all cops in addition to the current setup, ensuring confirmation of both intended behaviour when it comes to expected results and also the functionality of those expected results in Crystal.