Add expect_match_crystal spec helper
The split between rspec/ for testing that cops behave as expected, and test/ for testing that the corrected code has the expected behavior has always been an interim solution, and a flawed one at that. Having to test everything twice leads to code duplication, as well as some behavior going untested, and the test/ tests are impractical to run locally. With this, we can gradually transition to testing everything using RSpec, and sunset the test/ directory.

This helper doesn't yet work for everything covered by test/, and it also seems to be quite a bit slower, but we can fix those issues as they become relevant.